Split all services menu into multiple components.
