+++ src/components/AllServicesDropdown/common.ts
@@ -1,0 +1,2 @@
+export const TAB_CONTENT_ID = 'refTab1Section';
+export const FAVORITE_TAB_ID = 'favorites';
